implement lazy-loading VCS handler
When the VCS handler was eagerly loaded at the start of the plugin, it
would almost certainly load faster than the Git integration plugin for
VSCode would load the repositories, so the handler chosen would be a
standard file system one, incapable of tracking changes. This has been
fixed to only lazy-load the handler upon an actual conversion request.