fix rename not being staged properly
Previously, to stage the rename, repo.add was being called upon the new
.kt file, with a vscode.commands.executeCommand call on the old
main.java file to stage the removal. The latter however would fail
silently and not actually stage a deletion. However, since repo.add also
stages deletions, the replacement has been to use repo.add instead for
the deleted file on the condition that it was previously tracked.