fix issues with incorrect scoring
The scores generated by the analytics script would incorrectly weight
scores with the wrong operators (previously it was (passing * total) /
loc), which is incorrect - it should be (passing / total) * loc to
correctly weight files with bigger LOC, and count tests passed as
contributing to the file more. This has been fixed, and the final
normalised score sits within a range of [0,1] now.