fix race condition on accept/cancel action
Opening a new editor causes the shared currently active editor variable
to be changed, and so after running the commit action the user may have
changed the active editor variable causing the wrong tab to be closed.
This has been fixed by storing the currently open editor immediately on
accept and cancel.