create idempotent opening of diff previews
Previously, the diff preview would open a new tab each time you clicked
on it. This became an issue especially after the code programmatically
started opening the next conversion to be reviewed - the tree view
wouldn't have the converter selected, and when selected, 2 separate
conversion tabs would open. This is no longer the case.